refactor(network): extract native transport to client::native submodule
